+++ test/FakeXMLHttpRequest.js
